# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de manipulação de produtos",
    "description": "Essa API tem como objetivo a manipulaçação de produtos",
    "termsOfService": "http://localhost:3333/terms",
    "contact": {
      "email": "anderson.imobpower@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3333/v1",
      "description": "API de test"
    }
  ],
  "paths": {
    "/products": {
      "post": {
        "summary": "Cadastro de produtos",
        "description": "Essa rota será responsável por cadastrar um novo produto",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "examples": {
                "product": {
                  "value": {
                    "name": "Mouse USB",
                    "description": "Mouse USB RGB",
                    "price": 300
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Product Already Exists"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/products/findByName": {
      "get": {
        "description": "Busca de um produto pelo nome",
        "summary": "Busca de um produto pelo nome",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Nome do Produto para busca",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "description": "Busca de um produto pelo id",
        "summary": "Busca de um produto pelo id",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Nome do Produto pelo ID",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Alterar um produto pelo id",
        "summary": "Alterar um produto pelo id",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Nome do Produto pelo ID",
            "required": true
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "examples": {
                "product": {
                  "value": {
                    "name": "Mouse USB",
                    "description": "Mouse USB Sem fio",
                    "price": 1300
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Deletar um produto pelo id",
        "summary": "Deletar um produto pelo id",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Nome do Produto pelo ID",
            "required": true
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    }
  },

  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "id": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}